refactor to fix code smell
